Update client code (not finished) and projects
